+++ pyzoo/zoo/ray/data/dataset.py
@@ -1,4 +1,21 @@
+#
+# Copyright 2018 Analytics Zoo Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
 import numpy as np
+
 import tensorflow as tf
 
 class RayDataSet(object):
@@ -10,7 +27,10 @@ class RayDataSet(object):
 
     @staticmethod
     def from_dataset_generator(input_fn, batch_size, repeat=True, shuffle=True):
-        return TFDataSetWrapper(input_fn=input_fn, batch_size=batch_size, repeat=repeat, shuffle=shuffle)
+        return TFDataSetWrapper(input_fn=input_fn,
+                                batch_size=batch_size,
+                                repeat=repeat,
+                                shuffle=shuffle)
 
 # TODO: haven't finished yet
 class ArrayLikeDataset(RayDataSet):
